requirements.txt
browserstack-local >= 1.2.3
selenium~=4.27.1
browserstack-sdk
requests~=2.32.3
PyYAML~=6.0.2